🔨 Refactor: 쿠키 관련 유틸 함수 분리
